fix: preferences

+++ app/src/main/java/com/mishaismenska/hackatonrsschoolapp/data/UnitConverter.kt
@@ -1,5 +1,5 @@
 package com.mishaismenska.hackatonrsschoolapp.data
 
-fun kgToLb(kgs: Int) = kgs*2.2
-fun lbToKg(lbs: Int) = lbs/2.2
+fun kgToLb(kgs: Int) = (kgs*2.2).toInt()
+fun lbToKg(lbs: Int) = (lbs/2.2).toInt()
 fun mlToOz(mls: Int) = mls*0.03
